Added the Counter Tutorial.
Fixes #6

Added an LFE version of the counter tutorial from the blog post linked
in the issue description.

I tried to add some more interesting elements towards the end with
returning closures to avoid bouncing variables around etc. Not really my
forte but thought it would be fun to try.